src/property_prediction/requirements.txt
# Core scientific computing
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Machine learning and deep learning
torch>=1.9.0
torch-geometric>=2.0.0
scikit-learn>=1.0.0
xgboost>=1.5.0

# Molecular representation and cheminformatics
rdkit-pypi>=2021.9.1
mordred>=1.2.0

# MoleculeNet and related packages
deepchem>=2.7.0


# Data processing and visualization
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0

# Utilities
tqdm>=4.62.0
joblib>=1.1.0
pickle5>=0.0.11

# Optional: For advanced molecular featurization
dgl>=0.9.0
dgl-lifesci>=0.3.0 